cleanup and move all redundant d3d state definitions
