+++ ENIGMAsystem/SHELL/Graphics_Systems/OpenGL1/GLd3d.cpp
@@ -36,11 +36,6 @@ namespace enigma {
 
 void d3d_light_update_positions(); // forward declare
 
-bool d3dMode = false;
-bool d3dHidden = false;
-bool d3dZWriteEnable = true;
-int d3dCulling = 0;
-
 void graphics_set_matrix(int type) {
   enigma_user::draw_batch_flush(enigma_user::batch_flush_deferred);
   glm::mat4 matrix;
@@ -246,19 +241,6 @@ void d3d_set_culling(int mode)
   if (mode > 0){
     glFrontFace(windingstates[mode-1]);
   }
-}
-
-bool d3d_get_mode()
-{
-  return enigma::d3dMode;
-}
-
-bool d3d_get_hidden() {
-  return enigma::d3dHidden;
-}
-
-int d3d_get_culling() {
-  return enigma::d3dCulling;
 }
 
 void d3d_set_fill_mode(int fill)
